Iron Session 얻는 코드 분리
